refactor: optimize recipe
