+++ apps/mobile/app/(app)/(recipe)/_layout.tsx
@@ -3,13 +3,18 @@ import { Stack } from 'expo-router';
 export default function MenuLayout() {
   return (
     <Stack
-      initialRouteName="index"
       screenOptions={{
+        animation: 'slide_from_right',
         headerShown: false
       }}
     >
-      <Stack.Screen name="index" />
-      <Stack.Screen name="recipe" options={{ presentation: 'modal' }} />
+      <Stack.Screen
+        name="create-recipe"
+        options={{
+          presentation: 'modal',
+          title: ''
+        }}
+      />
     </Stack>
   );
 }
